backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.10.0
aiofiles==23.2.1
boto3==1.29.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
authlib==1.2.1
httpx==0.25.2
redis==5.0.1
alembic==1.13.0

